requirements.txt
Django>=3.2.13,<3.3.0
djangorestframework>=3.13.1,<3.14.0

flake8>=4.0.1,<4.1.0